requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
nba_api==1.4.1
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0